First pass at decode history

+++ internal/ui/ui.go
@@ -29,8 +29,8 @@ type MainWindow struct {
 	requestContent *PacketDisplay
 	// responseContent the response content of the currently selected packet
 	responseContent *PacketDisplay
-	// recordButton is the button that starts/stops recording when pressed
-	recordButton *RecordButton
+	// analysisToolbar is the top-level toolbar
+	analysisToolbar *AnalysisToolbar
 	// PacketFilter manages the list of packets currently loaded into GITM
 	PacketFilter *PacketFilter
 }
@@ -94,35 +94,61 @@ func MakeMainWindow(packetChan chan packet.Packet, restart func()) *MainWindow {
 	w := util.NewWindowIfNotExists(lang.L("GITM"))
 	filter := NewPacketFilter(w)
 	w.SetMaster()
+	decodeHistoryData := []string{}
+
+	var l *widget.List
+	l = widget.NewList(
+		func() int { return len(decodeHistoryData) },
+		func() fyne.CanvasObject {
+			return NewHistoryItem("History item 0")
+		},
+		func(lii widget.ListItemID, co fyne.CanvasObject) {
+			co.(*HistoryItem).SetText(decodeHistoryData[lii])
+
+			l.SetItemHeight(lii, co.(*HistoryItem).MinSize().Height)
+		},
+	)
+	handleDecodeResult := func(result string) {
+		decodeHistoryData = append(decodeHistoryData, result)
+		l.Refresh()
+	}
+
+	l.Hide()
+	// TODO: Fix the ugly double pointer situation with content
+	var content *container.Split
 	mainWindow := &MainWindow{
 		Window:          w,
-		requestContent:  NewPacketDisplay(lang.L("Request"), w),
-		responseContent: NewPacketDisplay(lang.L("Response"), w),
-		recordButton:    NewRecordButton(filter, w),
+		requestContent:  NewPacketDisplay(lang.L("Request"), w, handleDecodeResult),
+		responseContent: NewPacketDisplay(lang.L("Response"), w, handleDecodeResult),
+		analysisToolbar: NewAnalysisToolbar(filter, w, l, &content),
 		PacketFilter:    filter,
 		packetChan:      packetChan,
 	}
 	mainWindow.registerShortcuts(restart)
 
 	mainWindow.makeMenu(func() { settings.MakeSettingsUi(w, restart).Show() })
-
-	mainWindow.SetContent(
+	content = container.NewHSplit(
 		container.NewVSplit(
-			container.NewBorder(
-				container.NewVBox(
-					mainWindow.recordButton,
-					mainWindow.PacketFilter,
-					widget.NewSeparator(),
-				),
-				nil,
-				nil,
-				nil,
-				NewPacketList(mainWindow.PacketFilter, mainWindow),
-			),
+			NewPacketList(mainWindow.PacketFilter, mainWindow),
 			container.NewHSplit(
 				mainWindow.requestContent,
 				mainWindow.responseContent,
 			),
+		),
+		l,
+	)
+	content.SetOffset(.80)
+	mainWindow.SetContent(
+		container.NewBorder(
+			container.NewVBox(
+				mainWindow.analysisToolbar,
+				mainWindow.PacketFilter,
+				widget.NewSeparator(),
+			),
+			nil,
+			nil,
+			nil,
+			content,
 		),
 	)
 
@@ -137,7 +163,7 @@ func (m *MainWindow) StartPacketHandler() {
 	go func() {
 		for {
 			p := <-m.packetChan
-			if m.recordButton.IsRecording {
+			if m.analysisToolbar.IsRecording {
 				existingPacket := m.PacketFilter.FindPacket(p)
 
 				if existingPacket != nil {
@@ -152,7 +178,7 @@ func (m *MainWindow) StartPacketHandler() {
 
 // CheckForCrashData checks to see if there is data from a prior crash.
 //
-// If there is, it displays a comfirmation dialog for the user to choose whether to load the
+// If there is, it displays a confirmation dialog for the user to choose whether to load the
 // crash data.
 func (m *MainWindow) CheckForCrashData() {
 	configDir, err := util.GetConfigDir()
@@ -161,13 +187,13 @@ func (m *MainWindow) CheckForCrashData() {
 	}
 	crashData := configDir + string(os.PathSeparator) + "crash.json"
 	if _, err := os.Stat(crashData); err == nil {
-		dialog.NewConfirm(lang.L("Crash data found"), lang.L("There was crash data found. Want to load it?"), func(b bool) {
+		dialog.ShowConfirm(lang.L("Crash data found"), lang.L("There was crash data found. Want to load it?"), func(b bool) {
 			if b {
 				m.PacketFilter.LoadPacketsFromFile(crashData)
 			}
 			if err := os.Remove(crashData); err != nil {
 				util.ReportUiError(err, m)
 			}
-		}, m).Show()
+		}, m)
 	}
 }
